Extract `AssetManagerTest` asset bundle stubs into `tests/data/web`
