+++ tests/data/web/AssetManagerTestSourceBundle.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * @link https://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license https://www.yiiframework.com/license/
+ */
+
+namespace yiiunit\data\web;
+
+use yii\web\AssetBundle;
+
+class AssetManagerTestSourceBundle extends AssetBundle
+{
+    public $sourcePath = '@testSourcePath';
+    public $js = [
+        'js/jquery.js',
+    ];
+    public $css = [
+        'css/stub.css',
+    ];
+}
